+++ RELEASES.md
@@ -109,6 +109,7 @@ Upcoming
 - Bugfix 'Browse a Jdbc table' for 'MS SQL for WebSphere'
 - Add facility to generate flow images by CommandSender next to HttpSender
 - Bugfix "facility to hide properties" caused by splitting configuration
+- Introduction of CounterSwitchPipe (for load balancing between pipes)